Yet Another Coronavirus Question
Input file: standard input
Output file: standard output
Time limit: 2 seconds
Memory limit: 256 megabytes
Aladdin is the Supreme Leader of his country, Wadiya and wants to determine the epicentre of coronavirus
in his nation. Wadiya has n cities which are connected to each other by n 1 roads which are bidirectional.
The length of each road is 1 unit. For some weird reason, the virus can travel only along these roads.
One can travel from any city u to any city v through these roads.
There are h cities which are infected and are coronavirus hotspots. This list of cities may be
incomplete (i.e., there may be other cities which are also infected). Aladdin knows that the virus starts
spreading from the epicentre and reaches only those cities which are at a distance of atmost x units from
the epicentre. You are given the list of infected cities, and you have to determine the number of possible
epicentres of the virus. Note that it may be possible that some opposer to the regime corrupted the data,
and hence there is no valid epicentre that can be deduced from the data.